third_party/blink/web_tests/platform/generic/fast/canvas/OffscreenCanvas-transferToImageBitmap-expected.txt

PASS aCanvas.transferToImageBitmap() threw exception InvalidStateError: Failed to execute 'transferToImageBitmap' on 'OffscreenCanvas': Cannot transfer an ImageBitmap from an OffscreenCanvas with no context.
PASS image.width is 60
PASS image.height is 40
PASS ctx.lineWidth is 5
PASS imgData[0] is 255
PASS imgData[1] is 0
PASS imgData[2] is 0
PASS imgData[3] is 255
PASS imgData[0] is 0
PASS imgData[1] is 0
PASS imgData[2] is 0
PASS imgData[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
